Create Main module for Brittany
