+++ brittany.cabal
@@ -1,3 +1,4 @@
+cabal-version:        2.2
 name:                 brittany
 version:              0.12.1.1
 synopsis:             Haskell source code formatter
@@ -8,7 +9,7 @@ description: {
   .
   The implementation is documented in more detail <https://github.com/lspitzner/brittany/blob/master/doc/implementation/index.md here>.
 }
-license:              AGPL-3
+license:              AGPL-3.0-only
 license-file:         LICENSE
 author:               Lennart Spitzner
 maintainer:           Lennart Spitzner <hexagoxel@hexagoxel.de>
@@ -16,7 +17,6 @@ copyright:            Copyright (C) 2016-2019 Lennart Spitzner
                       Copyright (C) 2019 PRODA LTD
 category:             Language
 build-type:           Simple
-cabal-version:        1.18
 homepage:             https://github.com/lspitzner/brittany/
 bug-reports:          https://github.com/lspitzner/brittany/issues
 extra-doc-files: {
@@ -53,6 +53,7 @@ library {
     srcinc
   exposed-modules: {
     Language.Haskell.Brittany
+    Language.Haskell.Brittany.Main
     Language.Haskell.Brittany.Internal
     Language.Haskell.Brittany.Internal.Prelude
     Language.Haskell.Brittany.Internal.PreludeUtils
@@ -145,60 +146,12 @@ executable brittany
     buildable: True
   }
   main-is:             Main.hs
-  other-modules: {
-    Paths_brittany
-  }
-  -- other-extensions:
-  build-depends:
-    { brittany
-    , base
-    , ghc
-    , ghc-paths
-    , ghc-exactprint
-    , transformers
-    , containers
-    , mtl
-    , text
-    , multistate
-    , syb
-    , data-tree-print
-    , pretty
-    , bytestring
-    , directory
-    , butcher
-    , yaml
-    , aeson
-    , extra
-    , uniplate
-    , strict
-    , monad-memo
-    , unsafe
-    , safe
-    , deepseq
-    , semigroups
-    , cmdargs
-    , czipwith
-    , ghc-boot-th
-    , filepath >=1.4.1.0 && <1.5
-    }
   hs-source-dirs:      src-brittany
-  include-dirs:        srcinc
+  build-depends:
+    { base
+    , brittany
+    }
   default-language:    Haskell2010
-  default-extensions: {
-    CPP
-
-    NoImplicitPrelude
-
-    GADTs
-
-    FlexibleContexts
-    FlexibleInstances
-    ScopedTypeVariables
-    MonadComprehensions
-    LambdaCase
-    MultiWayIf
-    KindSignatures
-  }
   ghc-options: {
     -Wall
     -fno-spec-constr
